move launch detail windows out of rocket module
