+++ src/modules/window/building_detail_window.cpp
@@ -1,10 +1,10 @@
 #include "building_detail_window.h"
 #include "imgui.h"
+#include "launch_detail_window.h"
+#include "launch_window.h"
 #include "modules/base/assert.h"
 #include "modules/base/base.h"
 #include "modules/engine/gui.h"
-#include "modules/rocket/launch_detail_window.h"
-#include "modules/rocket/launch_window.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/site/rocket_prefab_window.h"
 #include "modules/site/site.h"
