+++ src/modules/window/contract_detail_window.cpp
@@ -1,10 +1,10 @@
 #include "contract_detail_window.h"
 #include "imgui.h"
+#include "launch_detail_window.h"
+#include "launch_window.h"
 #include "modules/engine/gui.h"
 #include "modules/engine/helpers.h"
 #include "modules/rocket/contract_actions.h"
-#include "modules/rocket/launch_detail_window.h"
-#include "modules/rocket/launch_window.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/window/contracts_window.h"
 #include "widgets/widgets.h"
