+++ src/modules/window/launch_detail_window.h
