+++ src/modules/window/rocket_detail_window.cpp
@@ -1,11 +1,11 @@
 #include "rocket_detail_window.h"
 #include "building_detail_window.h"
 #include "imgui.h"
+#include "launch_detail_window.h"
+#include "launch_window.h"
 #include "modules/base/base.h"
 #include "modules/engine/gui.h"
 #include "modules/engine/helpers.h"
-#include "modules/rocket/launch_detail_window.h"
-#include "modules/rocket/launch_window.h"
 #include "modules/rocket/rocket_actions.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/site/site.h"
